Added base Input component from Shadcn/ui library and extended it to a component named InputExtended. Updated occurences of old Input component with new InputExtended component. Deleted app/_components folder since there is no more need for that directory.
